FileSystem - pripraveni syscallu a propojeni s uzivatelskym rozhranim (rtl.cpp) + upraveni FS zapisovaci metody

+++ os_simulator/src/kernel/Thread_managment.cpp
@@ -11,7 +11,7 @@ void handleThread(CONTEXT &regs) {
 		regs.Rax = (decltype(regs.Rax))i;
 	}break;
 	case scExecuteThread: {
-		int id = regs.Rdx;
+		int id = (int)regs.Rdx;
 		int i = (decltype(regs.Rax)) tcb->execute_thread(id);
 	}break;
